requirements.txt
# Discord関連
discord.py==2.3.2
discord-ext-menus>=1.0.0a31
# discord-timestamp>=1.0.0 - パッケージが見つからないためコメントアウト

# データベース関連
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.28.0
redis==5.0.1
aioredis==2.0.1

# Webダッシュボード関連
fastapi==0.104.1
uvicorn==0.23.2
pydantic>=1.10.7
python-jose>=3.3.0
python-multipart==0.0.6
httpx>=0.24.0
jinja2>=3.1.2

# AI関連
google-generativeai==0.3.1
# openai>=0.27.0
tiktoken>=0.4.0
numpy>=1.24.0
scikit-learn>=1.2.0

# ユーティリティ
python-dotenv==1.0.0
pyyaml>=6.0
loguru>=0.7.0
humanize>=4.6.0
pytz>=2023.3
pillow==10.1.0
colorama>=0.4.6
aiohttp==3.9.1
ujson>=5.7.0
pyjwt==2.8.0
passlib==1.7.4
bcrypt==4.0.1 